- moved Descriptor.__set__ to C    - changed __set__ to not set if value unchanged, 650 times faster    - changed a bunch of calls to no longer use keyword args, faster    - released chandlerdb-0.5-4
git-svn-id: http://svn.osafoundation.org/chandler/trunk@4932 3c49585b-f0f7-0310-b5f9-dfe92a88fbfe

+++ chandler/repository/remote/RemoteFilter.py
@@ -128,7 +128,7 @@ class RemoteFilter(XMLFilter):
     def itemStart(self, attrs):
 
         self.itemUUID = UUID(attrs['uuid'])
-        self.itemVersion = long(attrs['version'])
+        self.itemVersion = int(attrs['version'])
 
         version = self.store.getItemVersion(self.itemVersion, self.itemUUID)
         if not self.force and version == self.itemVersion:
@@ -241,7 +241,7 @@ class RemoteFilter(XMLFilter):
             return
         
         uuid = UUID(attrs['uuid'])
-        version = long(attrs['version'])
+        version = int(attrs['version'])
         encoding = attrs['encoding']
         compression = attrs.get('compression', None)
         store = self.store
